requirements.txt
streamlit==1.31.0
google-generativeai==0.8.3
beautifulsoup4==4.12.3
requests==2.31.0
plotly==5.18.0
pandas==2.2.0
vaderSentiment==3.3.2
python-dotenv==1.0.1
lxml==5.1.0
